refactor : Change MapContainer name to MapController
